Fix language parameter is ignored in GraphQL response
