+++ CHANGELOG.adoc
@@ -10,6 +10,10 @@
 
 
 = Gentics Mesh Changelog
+[[v0.17.4]]
+== 0.17.4 (TBD)
+
+icon:check[] GraphQL: Fixed language parameter in nodes query method was ignored in some cases. link:https://github.com/gentics/mesh/issues/365[#365]
 
 [[v0.17.3]]
 == 0.17.3 (15.03.2018)
