+++ core/src/test/java/com/gentics/mesh/core/graphql/GraphQLNodeLanguageSearchEnpointTest.java
@@ -1,5 +1,6 @@
 package com.gentics.mesh.core.graphql;
 
+
 import com.gentics.mesh.FieldUtil;
 import com.gentics.mesh.core.rest.node.NodeCreateRequest;
 import com.gentics.mesh.core.rest.schema.impl.SchemaReferenceImpl;
@@ -10,41 +11,38 @@ import com.gentics.mesh.test.context.MeshTestSetting;
 import org.junit.Before;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
-import org.junit.runners.Parameterized.Parameters;
 
-import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 
 import static com.gentics.mesh.test.ClientHelper.call;
 import static com.gentics.mesh.test.TestDataProvider.PROJECT_NAME;
 
-@MeshTestSetting(useElasticsearch = true, testSize = TestSize.FULL, startServer = true)
 @RunWith(Parameterized.class)
-public class GraphQLNodeScrollSearchEndpointTest extends AbstractGraphQLSearchEndpointTest {
-
-	@Parameters(name = "query={0}")
+@MeshTestSetting(useElasticsearch = true, testSize = TestSize.FULL, startServer = true)
+public class GraphQLNodeLanguageSearchEnpointTest extends AbstractGraphQLSearchEndpointTest {
+	@Parameterized.Parameters(name = "query={0}")
 	public static List<String> paramData() {
-		List<String> testQueries = new ArrayList<>();
-		testQueries.add("node-elasticsearch-scroll-query");
-		return testQueries;
+		return Collections.singletonList("node-elasticsearch-language-query");
 	}
 
-	public GraphQLNodeScrollSearchEndpointTest(String queryName) {
+	public GraphQLNodeLanguageSearchEnpointTest(String queryName) {
 		super(queryName);
 	}
 
 	@Before
 	public void createNodes() {
 		String parentNodeUuid = tx(() -> folder("2015").getUuid());
-		for (int i = 0; i < 100; i++) {
+		Arrays.asList("de", "en").forEach(lang -> {
 			NodeCreateRequest nodeCreateRequest = new NodeCreateRequest();
 			nodeCreateRequest.setParentNode(new NodeReference().setUuid(parentNodeUuid));
 			nodeCreateRequest.setSchema(new SchemaReferenceImpl().setName("content"));
-			nodeCreateRequest.setLanguage("en");
-			nodeCreateRequest.getFields().put("slug", FieldUtil.createStringField("test" + i));
+			nodeCreateRequest.setLanguage(lang);
+			nodeCreateRequest.getFields().put("slug", FieldUtil.createStringField("test" + lang));
 			nodeCreateRequest.getFields().put("teaser", FieldUtil.createStringField("some teaser"));
 			nodeCreateRequest.getFields().put("content", FieldUtil.createStringField("Blessed mealtime again!"));
-			call(() -> client().createNode(PROJECT_NAME, nodeCreateRequest, new NodeParametersImpl().setLanguages("en")));
-		}
+			call(() -> client().createNode(PROJECT_NAME, nodeCreateRequest, new NodeParametersImpl().setLanguages(lang)));
+		});
 	}
 }
