+++ core/src/test/java/com/gentics/mesh/core/graphql/GraphQLUserScrollSearchEndpointTest.java
@@ -1,29 +1,19 @@
 package com.gentics.mesh.core.graphql;
 
-import static com.gentics.mesh.assertj.MeshAssertions.assertThat;
-import static com.gentics.mesh.test.ClientHelper.call;
-import static com.gentics.mesh.test.TestDataProvider.PROJECT_NAME;
-
-import java.util.ArrayList;
-import java.util.List;
-
+import com.gentics.mesh.test.TestSize;
+import com.gentics.mesh.test.context.MeshTestSetting;
+import com.syncleus.ferma.tx.Tx;
 import org.junit.Before;
-import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import org.junit.runners.Parameterized.Parameters;
 
-import com.gentics.mesh.core.rest.graphql.GraphQLResponse;
-import com.gentics.mesh.test.TestSize;
-import com.gentics.mesh.test.context.AbstractMeshTest;
-import com.gentics.mesh.test.context.MeshTestSetting;
-import com.syncleus.ferma.tx.Tx;
-
-import io.vertx.core.json.JsonObject;
+import java.util.ArrayList;
+import java.util.List;
 
 @MeshTestSetting(useElasticsearch = true, testSize = TestSize.FULL, startServer = true)
 @RunWith(Parameterized.class)
-public class GraphQLUserScrollSearchEndpointTest extends AbstractMeshTest {
+public class GraphQLUserScrollSearchEndpointTest extends AbstractGraphQLSearchEndpointTest {
 
 	@Parameters(name = "query={0}")
 	public static List<String> paramData() {
@@ -32,10 +22,9 @@ public class GraphQLUserScrollSearchEndpointTest extends AbstractMeshTest {
 		return testQueries;
 	}
 
-	private final String queryName;
 
 	public GraphQLUserScrollSearchEndpointTest(String queryName) {
-		this.queryName = queryName;
+		super(queryName);
 	}
 
 	@Before
@@ -46,17 +35,6 @@ public class GraphQLUserScrollSearchEndpointTest extends AbstractMeshTest {
 				createUser(username + i);
 			}
 		}
-	}
-
-	@Test
-	public void testNodeQuery() throws Exception {
-		try (Tx tx = tx()) {
-			recreateIndices();
-		}
-		GraphQLResponse response = call(() -> client().graphqlQuery(PROJECT_NAME, getGraphQLQuery(queryName)));
-		JsonObject json = new JsonObject(response.toJson());
-		System.out.println(json.encodePrettily());
-		assertThat(json).compliesToAssertions(queryName);
 	}
 
 }
